PHP 8.1 warning issue
Hi gents,
it looks like we get warning in PHP 8.1 on line 767 if $val is null (may happen if not all values passed in). So I would add a line to check if $val is set and if not, reset it to empty string. This avoids the warning and keeps the functionality as is.
Cheers, David